+++ nodes/__init__.py
@@ -1,6 +1,6 @@
 """HomeKit Hub Node Server nodes."""
 
-VERSION = "0.1.14"
+VERSION = "0.2.0"
 from .Controller import Controller as Controller  # noqa: E402,F401
 
 __all__ = ["Controller", "VERSION"]